fix(roc-plugin-repo): Support for an ALL, MULTI and * scope, for mono
ALL will make the commit be applied to all projects, for example when changing the build configuration. Can be in lowercase.

MULTI enables support for listing the scopes in the footer of the commit, using either SCOPES or SCOPE (SCOPE: roc-plugin-repo) in a similar fashion to breaking changes. Can be in lowercase.

* is automatic scope detection and will be applied to the changed directories, will not work for the all scope.